move modules around some more and fix names to get tests passing again
